+++ Js2IL.Tests/Hosting/JavaScript/nestedReturn.js
@@ -1,0 +1,18 @@
+"use strict";
+
+exports.getWindow = function () {
+  return {
+    document: {
+      title: "Hello"
+    }
+  };
+};
+
+exports.getTitleViaHost = function () {
+  // Useful for sanity: this should work in-script too.
+  return exports.getWindow().document.title;
+};
+
+exports.getTitle = function (win) {
+  return win.document.title;
+};